2021-05-09/写笔记带后端/library-server-develop/requirements.txt
attrs==16.3.0
beautifulsoup4==4.5.3
cffi==1.9.1
chardet==2.3.0
constantly==15.1.0
cryptography==1.7.1
cssselect==1.0.1
Django==1.10.5
html5lib==0.999999999
idna==2.2
incremental==16.10.1
lxml==3.7.2
parsel==1.1.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.17
pycrypto==2.6.1
PyDispatcher==2.0.5
PyJWT==1.4.2
PyMySQL==0.7.9
pyOpenSSL==16.2.0
python-weixin==0.1.6.1
pywin32==220
queuelib==1.4.2
requests==2.13.0
Scrapy==1.3.0
service-identity==16.0.0
simplejson==3.10.0
six==1.10.0
Twisted==16.6.0
w3lib==1.16.0
webencodings==0.5
zope.interface==4.3.3
